+++ packages/react-magma-dom/src/components/Pagination/SimplePagination.test.js
@@ -65,6 +65,59 @@ describe('Simple Pagination', () => {
     expect(nextButton).not.toBeInTheDocument();
   });
 
+  it('should set the page number to the defaultPage', () => {
+    const { getByTestId } = render(
+      <Pagination
+        type={PaginationType.simple}
+        count={4}
+        testId={testId}
+        defaultPage={2}
+      />
+    );
+
+    expect(getByTestId(`${testId}-select`).value).toBe('2');
+  });
+
+  it('should set the page number to the page', () => {
+    const { getByTestId } = render(
+      <Pagination
+        type={PaginationType.simple}
+        count={4}
+        testId={testId}
+        page={2}
+      />
+    );
+
+    expect(getByTestId(`${testId}-select`).value).toBe('2');
+  });
+
+  it('should set the page number to 1 if the passed page is greater than the count', () => {
+    const { getByTestId } = render(
+      <Pagination
+        type={PaginationType.simple}
+        count={4}
+        testId={testId}
+        page={6}
+      />
+    );
+
+    expect(getByTestId(`${testId}-select`).value).toBe('1');
+  });
+
+  it('should set the page number to the page when both page and defaultPage are passed', () => {
+    const { getByTestId } = render(
+      <Pagination
+        type={PaginationType.simple}
+        count={4}
+        testId={testId}
+        defaultPage={2}
+        page={3}
+      />
+    );
+
+    expect(getByTestId(`${testId}-select`).value).toBe('3');
+  });
+
   describe('Disabled', () => {
     it('Should render disabled navigation icons, text, and select', () => {
       const { getByText, getByLabelText } = render(
@@ -184,7 +237,8 @@ describe('Simple Pagination', () => {
         <Pagination
           type={PaginationType.simple}
           count={3}
-          defaultPage={2}
+          defaultPage={1}
+          page={2}
           testId={testId}
         />
       );
@@ -329,7 +383,7 @@ describe('Simple Pagination', () => {
       <Pagination
         type={PaginationType.simple}
         count={4}
-        defaultPage={2}
+        page={2}
         testId={testId}
         onPageChange={onClickMock}
       />
@@ -365,6 +419,7 @@ describe('Simple Pagination', () => {
       <Pagination
         type={PaginationType.simple}
         count={3}
+        page={3}
         testId={testId}
         onPageChange={onClickMock}
       />